requirements.txt
torch
langchain_teddynote
protobuf>=5.0,<6.0
pymupdf
langchain
langchain_community
langchain_openai
langchain_ollama
pypdf
chromadb
tiktoken
httpx==0.27.0
streamlit
faiss-gpu-cu12
matplotlib
langchain_anthropic
umap-learn
scikit-learn
transformers
langchain_chroma